// APP_UI/src/models/imprest.model.ts
export interface Imprest {
  serial_number?: any;
  id?: number | string;
  name: string;
  cameras?: string | any;
  serialNo:string;
  description: string;
  phone_number_1?: number | string;
  extension_1?: number | string;
  phone_number_2?: number | string;
  extension_2?: number | string;
  hospital_id?: string | Number | boolean;
  active?: boolean;
}